src/main/java/es/upm/grise/profundizacion2018/examenFinal/Greeting.java
package es.upm.grise.profundizacion2018.examenFinal;

public class Greeting {
	
	int hour;
	
	public int getHour() {
		return hour;
	}

	public void setHour(int hour) {
		this.hour = hour;
	}

	public String getGreeting( Language language ) {
		// Get current hour
		this.setHour(new MyCalendar().getHour());
		
		// Find out the greeting language
		if ( language == null )
			language = Message.getDefaultLanguage();
		
		// Get the moment of the day
		TimeOfTheDay moment;
		if( hour < 12 ) 
			moment = TimeOfTheDay.MORNING;
		else if ( hour < 18 )
			moment = TimeOfTheDay.AFTERNOON;
		else
			moment = TimeOfTheDay.EVENING;
			
		// Return the message		
		return Message.getMessage( moment, language );
	}

}
